Remove static var from receiver, move ConnectivityReceiver to main package
