+++ app/src/main/AndroidManifest.xml
@@ -108,7 +108,7 @@
 
         <!-- The below code is for android OS version below N -->
         <receiver
-            android:name=".common.ConnectivityReceiver"
+            android:name=".features.main.ConnectivityReceiver"
             android:enabled="true">
             <intent-filter>
                 <action android:name="android.net.conn.CONNECTIVITY_CHANGE" />
